Make the timeline chat a group chat, enabling (so much better) support for multiple accounts. One group chat is created per account, and messaging that group chat will send an update.
The character counting code needs to be updated to support counting in group chats, since right now it checks the AIChat's listObject, which seems wrong.

A few minor changes included: only invalidate the password if we get a 401 from Twitter, always update the timeline (open a chat if necessary) when we get an update, fix a crash when we try and invalidate an already-invalidated timer, pull 100 tweets (10 for user info) at a time.