Add a bit.ly service for URL shortening.
This currently uses an "adiumofficial" account as the account it's tied to. History is not enabled (history=1 would have to be appended to the link), so other than being a stupid requirement, users won't be able to claim this URL. :/ I'm not writing a UI for setting the username/API key so whatever.