Avoid the crash without fixing the problem. Refs #13172.
This is being caused by the AdiumApplescriptRunner telling GBApplescriptFiltersPlugin that an applescript finished too many times.

AdiumApplescriptRunner is being told by the ApplescriptRunner that it's finishing too many times. When it occurs before the dict removal, it ends up telling the filter it finished too many times; this is still race-conditiony broken, so this is just a hacky bandaid until I can sort out why it's happening.

I'm not sure why the ApplescriptRunner is doing so.