Implementation of ticket #14354: Contact list search should treat whitespace-separated words as keywords
Instead of searching contacts for the literal string, break it on word boundaries and search for contacts containing all of the keywords. Allows for quick searches, e.g. searching for Jonathan Dovercourt is as simple as "Jo Do".

Fixed memory management critiques by xnyhps and put in an explanitory comment for the choice of using CFStringTokenizer.